requirements.txt
numpy
scipy
matplotlib
sklearn
mne
pylsl
PyQt5
websocket-client==0.48.0
